Пишем первую функцию
