+++ index.php
@@ -332,25 +332,7 @@ $user_name = 'Павел'; // укажите здесь ваше имя
                 ]
             ];
             ?>
-            <?php
-            function cutText($post, $lenght)
-            {
-                $words = explode(" ", $post);
-                $sum = 0;
-                $words_back = array();
-                foreach ($words as $word) {
-                    $sum = mb_strlen($word) + $sum;
-                    if ($sum > $lenght) {
-                        $words_back[] = '...<a class="post-text__more-link" href="#">Читать далее</a>';
-                        break;
-                    } else {
-                        $words_back[] = $word;
-                    }
-                }
-                $post = implode(" ", $words_back);
-                return $post;
-            }
-            ?>
+            <?php require ('function.php')?>
             <?php foreach ($posts as $post): ?>
                 <article class="popular__post post <?= htmlspecialchars($post['type']) ?>">
                     <header class="post__header">
